Move `struct Registers` into own private module
